Drop file from index when replaced by directory
When the index is updated with a path that contains a directory that has
the same name as a previously added file, the file should be dropped
from the index.

For example, if the index contains "alice.txt", and we add
"alice.txt/bob.txt" (where "alice.txt" is now a directory), the *file*
"alice.txt" has to be removed from the index.