Recursively drop directory from index when replaced by file
When the index is updated with a file that has the same path as a
directory that was previously added, the directory and the entire tree
under it has to be removed from the index. To this end, the index
receives a new object @parents, which maps directory names to all the
files they contain, recursively.