Handle missing file permissions
When trying to add a file for which we don't have access permissions,
give up and exit with a non-zero exit status. Unlike for non-existing
files, the files listed before the unreadable one are getting added to
the object database, which doesn't immediately makes sense, but this is
how Git handles the same scenario.