Start detecting changes between index and workspace
The "jit status" command now also detects modifications between files in
the workspace and in the index. Changed content currently just compares
blob size; additionally, changed file modes are recognized.

The Rakefile is updated to require "standard/rake" instead of calling it
manually in a task.

Due to a strange false positive for escaped leading spaces in here-docs,
a standard config file is added to ignore that specific case for now.