Only hash files if timestamps between index and workspace differ
As an optimization for file change detection between index and
workspace, "jit status" now only hashes files to detect changes in
content if the timestamps between workspace and index disagree.

If it turns out that the contents are the same, the timestamps in the
index are updated to match the workspace, so next time, the same file
won't have to be hashed.