Fix file order for long status format
Files that exist in the head commit, but are deleted from the index are
the last thing we check for, and the initial implementation of the long
status format would always put them last, potentially resulting in the
filenames not being ordered alphabetically.

This change adds a sorted hash class, which overrides "each" to yield
its keys in order, using a sorted set under the hood. Changes to the
index and workspace use the new sorted hash to determine the status.