Add diffing between head and index
This adds a "--cached" option to the diff command, to compare the head
tree to the index. Unlike comparisons between working tree and index,
this also reports staged added files.

Like for status tests, because we haven't implemented deleting files
from the index yet, tests for this delete the file in the working
directory, then blow away the entire index, and build it again from the
working directory.