Use hunks to show shorter diffs
Instead of showing diffs in their entirety, this filters out long
stretches of lines common to both files; regions with edits ("hunks")
are shown with a three line context window and a header each.

Hunks are implemented as a new class in the Diff mobule, and edits now
store both lines.

Tests are updated to expect the full diff, and some unit tests for hunks
are added.

Test classes don't subclass each other any longer, as this resulted in
executing the tests in the superclass multiple times.