Fix flaky test for changed file with same size
When comparing a file between workspace and index, we first compare
metadata. If the size is unchanged, we compare mtime and ctime, and if
they differ, we compare the contents (and update the timestamps in the
index in any case). If they times don't differ, we assume that the file
hasn't been changed, without even looking at its contents.

The test for this would sometimes fail: a file was modified by changing
its contents, but keeping the size the same. Jit should notice the
difference in timestamps, compare the contents, and report the file as
modified. However, from time to time, it would fail to do so.

I ended up printing the stat values from the workspace file and the
index entry, for atime and ctime. When the test failed, that would look
something like this:

  1699139749 workspace ctime
  1699139749 index ctime

  3495849 workspace ctime nsec
  3495849 index ctime nsec

  1699139749 workspace mtime
  1699139749 index mtime

  3495849 workspace mtime nsec
  3495849 index mtime nsec

All mtime and ctime stamps were the same, down to the nanosecond! This
resulted in Jit assuming that there was no change. I don't believe it's
possible that commiting the file and then making a change to it can
happen in less than one nanosecond, but somehow the timestamps implied
that. (Something like a clock resolution maybe?)

To fix this, the test now sleeps for one millisecond before modifying
the file. This seems get rid of the occasional flakiness.