Add branch creation at given revision
This adds a revision parser that understands "@" (short for "HEAD"), the
parent of a revision "rev^", and an ancestor of a revision, "rev~5".
These shorthands can be nested, like in "@^^~3", which is "the third
ancestor of the parent of the parent of HEAD".

Revisions can be branch names, or commit IDs. IDs for objects that
aren't commits are handled.

Commit IDs can be shortened, as long as they are unambiguous. If more
than one ID in the object database matches a shortened ID, all
candidates are printed, together with their object type, and for commits
with the commit date, author, and commit message title.