Add functionality to diff trees
This adds a new class Database::TreeDiff that can produce a changes hash
from two input trees, enabling converting the first tree into the second
tree without having to build a tree from scratch.

It recursively looks for deletions (tree entries that exist in the
first, but not in the second tree), and then additions (tree entries
that exist only in the second tree). Directories are ignored, so a file
being replaced by a directory of the same name is considered deleted.