Implement checkout functionality to update workspace
This adds a new command, "checkout", which takes advantage of the tree
diffing functionality introduced previously. So far, checkout generates
the tree diff between the current tree-ish and the target tree-ish,
creates a migration for it, and applies it to the workspace. The index
remains unchanged for now, and we're not doing anything with the HEAD
ref.

The new Migration class converts a tree diff into a sets of changes,
taking care of deleting and creating directories at the right time so
they can't clash with files.

Actually applying the migration happens in the Workspace class.