Extract status comparison functions into inspector
To figure out if a checkout would overwrite uncommitted changes, we need
to know what's changing. To this end, we're extracting some status
functionality into a new Inspector class, with methods to indicate the
change a specific file has undergone, both for comparisons between index
and workspace, and commit tree and index.