Use symbolic references when checking out branches
When checking out a branch, we don't just write the object ID of the
commit at the tip of the branch being checked out into HEAD, but a
symref to the new branch instead.

Methods to read from a ref file now understand both object IDs and
symrefs, and resolve until they reach an object ID.

Checking out a relative reference, i.e., not a branch, still results in
"detached HEAD" with HEAD containing directly an object ID instead of a
symref.