Add options to decorate logs with refs
This addes the "--decorate[=<format>]" option to the log command.
Options are "auto", "short" (default), "full", or "no" ("--no-decorate).
"short" prints the short ref names, and "full" the full path to a ref.
If HEAD points to a ref, this is shows like "HEAD -> main".

To enable decorations, the log command populates a lookup table from
object IDs to refs, and for each commit in the log checks if there are
entries in that table.