Add option to print patches in logs
This extends the log options with "-p" (aliases: "--patch", "-u") to
print the diff of a commit to its parent in the log, and "-s" (alias:
"--no-patch") to suppress it. The options also apply to the diff
command, but the defaults are different: diff by default prints patches,
log by default does not.

Additionally, the diff command now also accepts two revisions and shows
the diff between them.