Handle logging multiple branches
When log receives multiple revisions as positional parameters, we handle
this by building a queue of commits sorted by author date (will be
committer date later, but so far they're always the same). The heads of
all branches are added to that queue, and when processing a commit (by
printing its logs), the parent gets added to the queue.

Flags on the @commits hash keep track of object IDs we've already seen,
so every commit gets processed and printed only once.