Allow exclusion of revisions
This teaches the RevList class two new revision patterns: "^rev" and
"rev1..rev2". A log command that looks like

  jit log branch1 branch2 ^branch3

means "list all commits reachable from branch1 and branch2, then remove
those reachable from branch 3"; think set operations.

The notation

  jit log branch1..branch2

is a shorthand for

  jit log branch2 ^branch1

Which, if branch1 is an ancestor of branch2, is the same as "all the
changes in branch2 since branch1".

In addition to parsing these new revision primitives, the RevList class
is updated to use an algorithm that stores commits to list in memory
first, so it can mark them as "uninteresting" if they're reachable from
a revision that should be excluded.