Allow filtering logs by files and directories
This modifies TreeDiff such that it accepts a PathFilter, and only
checks matchings paths; PathFilter is implemented as a trie.

If the "log" command receives additional positional parameters, they are
interpreted as file or directory names for which to filter the logs.
They are added to a prune list if they exist in the workspace, and tree
are marked as "treesame" is they don't change any of these files.
Treesame commits are removed from the results.