Allow commits to have multiple parents
To enable merge commits, this adds the option for a commit to have
multiple parents. The functionality to write a commit is extracted into
a new shared module, and commit writing and parsing are modified to
create and understand mulitple parent entries.

To keep commands that rely on the "parent" method working, we re-add it,
returning the first element of the parents array.