Add simple implementation of merge command
This adds a merge implementation that assumes merge-free history, and no
conflicts on merges. It takes a revision to be merged as its argument,
calculates the tree diff the merge base and that revision, and applies
the diff to HEAD. At the end, it creates a new commit with the old HEAD
as the first parent, and the merged commit as the second parent.