Find best common ancestor in merging histories
Instead of relying the history to be a tree, we now can also find the
best common ancestor in histories with merges. Instead of returning the
first commit that is reachable from both starting commits, we keep a
list of such commits, and mark them as stale if they're reachable from
each other.

We also follow both parents when traversing history, and not just the
first (and only) parent like before.

This functionality is layered into a new Bases class, which uses
CommonAncestors internally, but maintains the aforementioned list.

We can't yet handle the situation where there are multiple best common
ancestors, and return all of them for now.