Detect merge conflicts
When both parents have modified the same file, there might be a merge
conflict, specifically when

- both sides have modified an existing file
- one side has modified a file and the other side has deleted it
- both sides have added a file at the same path with different content

If object ID and mode of a conflicted file can be merged, the conflict
is resolved. If not, we write both version of the file to the workspace,
and write the conflict to the index by storing base commit and both
parents with non-zero stages.