Detect file-directory conflicts during merge
Sometimes, the two sides of a merge make changes that cannot be combined
into a single tree. For example, if one side creates a file f.txt, and
the other side creates a file f.txt/g.txt., i.e., creates a _directory_
named f.txt., there's no way to combine them into a single tree.

In this case, we keep the most deeply nested path in the index and the
workspace, and the outer path gets marked as conflicted. A copy of the
file at the outer path marked with its origin, for example "f.txt~HEAD",
is stored in the workspace as an untracked file.

This commit also renames the handle_merged_ancestors method to
handle_merged_ancestor.