Show conflicts in status
This adds a mapping from conflicted index entries to how conflicts
appear in the long and short status. The Repository::Status class stores
all conflicted index entries with their stage(s) in a sorted hash, and
the status command then maps those to the corresponding messages.

For example, if an entry has stages 1 and 2, that means that the base
branch modified it and the head branch deleted it. In the long status,
this shows up as

  $ jit status
  on branch main
  Unmerged paths:

          deleted by them: <file>

or

  $ jit status --porcelain
  UD <file>