Persist pending merge and resume after conflict
When starting a merge, we store the merge head and the commit message in
two special files, .git/MERGE_HEAD and .git/MERGE_MSG. If the merge
succeeds, they are deleted.

If the merge is conflicted, the files are used to retrieve merge head
and commit message when the conflict is resolved. We introduce "merge
--continue", and add some error logging when a merge fails. After a
conflict has been resolve, we can now also use "commit" to continue.
Both "merge --continue" and "commit" fail if the conflict is not
resolved.

"merge --continue" fails if there is no merge in progress, as indicated
by the absence of the MERGE_HEAD and MERGE_MSG files.