Add support for combining diffs
A merge commit can contain changes between itself and all its parents,
and a "combined diff" is a way to visualize that.

For example, if the diffs of the merge commit against its left and right
parents look like

    Left          Right
  -------------------------
   -alfa          echo
   +echo          bravo
    bravo        -charlie
    delta        +delta
   +foxtrot      +foxtrot

then the combined diff looks like

   - alfa
   + echo
     bravo
    -charlie
    +delta
   ++foxtrot

This is done in a new class that does the grouping by converting an
array of diffs (Edit objects) and creating an array of arrays of edits
(Row objects).