Bail on "rm" when there are changes
This adds functionality to first scan for any unstaged changes, or
staged changes that aren't committed. If there are any, "rm" exits with
an error message and doesn't delete any files.

Notice that the tests required a sleep to avoid flakiness: sometimes,
creating a file and then immediately modifying it would (weirdly) result
in all timestamps on the file in the workspace match the timestamps
stored in the index, and "status" would think of the file as unmodified.
This is similar to the fix in commit d5840c9.