Add "--cached" and "--force" options to "rm"
To remove a file only from the index, but not the workspace, "rm"
considers "--cached". This still fails if the index is different from
both the workspace and HEAD, to avoid data loss.

By adding "--force", this check can be bypassed. "--force" also allows
removing a file that has an unstaged change, which is then also lost.