Add "--hard" mode to reset
Hard mode behaves liked mixed, but additionally restores the workspace
to match the index.

Untracked files should be left alone, so we cannot just update the index
and then the workspace, as we'd lose track of what is not tracked.

It's also not efficient to rebuild the entire workspace from scratch, as
we have to read each blob from disk.

Instead, we determine which paths are different, and update them.