Open editor to edit commit messages
When no commit message is provided via "-m"/"--message" or "-F"/"--file"
flags, this opens an editor to compose the the message.

To do so, a new Editor class is introduced with an "edit" singleton
method, which takes a path, an environemnt hash, and a block. The given
file is opened for writing, and the block and do so using "puts" and
"note" (for comments). When the editor is closed, "edit" returns the
contents of the file.

This is used for the following cases (unless flags like "--message" or
"--no-edit" are used):

- When committing

- When merging successfully

- On a conflicted merge, to write a list of conflicted files in the
  comments

- When resuming from a merge conflict