Add cherry-pick command
This adds a first version of the cherry-pick command, capable of picking
a single commit into the current branch.

For this history

  A   B   C
  o---o---o [main]
       \
        o---o
        D   E

the command

  jit cherry-pick E

results in

  A   B   C   E'
  o---o---o---o [main]
       \
        o---o
        D   E

where E' is the result of applying the diff between D and E on top of C.
This is done using a three-way merge between C and E, using D as the
base.

The resulting commit has only one parent.