Support cherry-pick for pending commits
When a commit is pending, we store the merged commit ID in
.git/MERGE_HEAD to reuse for "commit --continue". We need something
similar for cherry-picking, but we can't use the exact same mechanics,
as the result of cherry-picking is a commit with just one parent.

To this end, this extends the PendingCommit class to handle different
merge types, and supports a new .git/CHERRY_PICK_HEAD file for pending
cherry-pick merges.