Allow selecting multiple commits and ranges for cherry-pick
Instead of allowing to specifiy just a single commit for cherry-picking,
this adds the ability to specify multiple commits and ranges, and
applies them in sequence.

To this end, the RevList class is slightly adjusted with a "walk"
option: if none of the arguments are ranges, we don't walk the revision
history. We also have to revert the order in which RevList yields
commits, as it was originally designed for "jit log", using reverse
chronological order.