Handle conflicts in a cherry-pick range
When processing multiple commits in cherry-pick, PendingCommit only
stores one commit in .git/CHERRY_PICK_HEAD; if a commit halfway through
a range fails, the rest of the range is dropped.

To address this, this introduces a new class Sequencer, which keeps
track of the entire list of commits to cherry-pick. When resuming with
"--continue" or committing after fixing a conflict, the list of commits
still to pick is loaded from a file at .git/sequencer/todo.