Abort and quit for cherry-pick sequences
This adds two options to use while cherry-picking:

- "--quit" removes cherry-pick state by deleting pending commit and
  sequencer state, but leaves everything else in place

- "--abort" restores everything to the way it was before starting
  cherry-picking

To support this, the Sequencer class uses two new files in
.git/sequencer: "abort-safety" and "head". "head" contains the original
HEAD, and "abort-safety", has the ID of the most recent successfully
picked commit. If HEAD doesn't match the contents of "abort-safety", the
state has been messed with, and we can't (and don't) safely hard-reset
to the original HEAD state.