Consider XDG base spec for config file stack
For the global config file, Git recognizes two locations: ~/.gitconfig,
and $XDG_CONFIG_HOME/git/config, where $XDG_CONFIG_HOME defaults to
~/.config.

If ~/.gitconfig exists, it is used as the global config file. If it
doesn't exist and the XDG config file exists, then the XDG file is used.
The two files aren't combined, either one or the other is used.

This change checks for the existence of the two files and sets the path
for the global config file accordingly.