Select mainline parent for cherry-pick and revert
Previously, when reverting or cherry-picking a merge commit, we always
used the first parent as the base or target of the merge.

In this commit, a "-m"/"--mainline" option is added to the cherry-pick
and revert commands, which allows selecting which parent to use as the
mainline. To persist this option in case of conflicts, it is stored in
.git/sequencer/opts, using the Git config file format introduced
previously.