Update RevList to find trees and blobs
When transferring objects from one repository to another, we want to
minimize the data sent: we only want to send objects that the receiver
doesn't yet have.

To determine what objects these are, the RevList class is extended to
not just list commit IDs, but build a list of trees and blobs to send.
We also want to minimize the number of these.

Trees of "uninteresting" commits are also uninteresting: we mark trees
of parents of interesting commits as such. After traversing all commits,
the trees are examined and uninteresting ones (and their blobs) are
excluded.