Implement the pkt-line protocol
To do things like collecting data to send to a client, Git runs programs
on the remote that communicate via stdin/stdout; the connection itself
uses SSH.

The protocol the communication follows is called the pkt-line protocol,
see https://git-scm.com/docs/protocol-common. The client implements
methods to send and receive a message, and receive messages until a
certain string appears.

Each message has a 4-byte header indicating the message length
(including the header itself and a newline at the end). A header of
"0000" is a "flush packet". Invalid headers are assumed to be handled by
the caller and just handed off to it.

The first message includes a list of capabilities, which we store.