Add capability to write packs
This adds a new module, Pack, with a Writer class. It implements the
pack-*.pack file format (see https://git-scm.com/docs/pack-format):
after a with the word PACK, a version (2 in this case), and the number
of objects in the pack, it is followed by all the objects. We don't
apply delta compression (yet).

Each object starts with a header with type and (uncompressed) size, then
the zlib-compressed data.

The writer is presented a list of object IDs, which it then pulls from
the database and writes to the provided output stream.

The object size uses variable-length little-endian integers, which are
implemented in the VarIntLE module.

The method to load objects from the database is refactored into two
methods, one of which returns the header, and the other parses the
contents. We only need the header, and then pass the contents through
"raw" (unparsed).