Add pack reader
This adds a Reader class to the Pack module to invert what the Writer
class does. It takes an IO stream and provides methods to read a pack
header and pack objects.

Objects are read in chunks using read_nonblock (so we don't block when
we hit the end of the pack) and seek (to correct for overfetching on an
object).

VarIntLE gets a read method so we can convert a variable-length byte
sequence back to a number.