Add Stream class to wrap IO
When reading packs, we sometimes have to be able to seek backwards.
That's no problem when reading from a file, but you can't seek backwards
in a stream (like when doing a fetch from a remote repository).

This adds Stream class that implements enough of the IO interface to
allow us to seek backwards in a stream by keeping the current object in
memory.