Support shorthand refspecs
When pushing, we often specify the refspec directly on the command line
instead of checking remote.<name>.push.

Command line refspecs support shorthands:

- A missing colon means that source and target are identical:

    main --> main:main

- Names can be abbreviated:

    heads/topic   --> refs/heads/topic
    remotes/topic --> refs/remotes/topic
    topic         --> refs/heads/topic

So, "main" is short for "refs/heads/main:refs/heads/main".

If a value isn't a valid ref name, it isn't prefixed; this supports
specs like "@~2:main".

If a refspec starts with ":", i.e., there is no source, it means that
<nil> should be copied into that ref, effectively deleting it.