Support compare-and-swap for ref updates
The push command sends ref updates in the format

  <old> <new> <ref>

for example

  1234abc 5678def refs/heads/main

This is to prevent multiple client from updating the same ref to
different values, which would result in losing updates. If the current
value of the ref isn't <old>, the update is rejecte; during an update,
the ref is locked.