Add push and receive-pack commands
Quite similar to fetch and upload-pack, this adds commands to send
objects to the remote repo, and updating "refs/heads" in the remote.

push calls receive-pack in the remote repo, gets all the refs from it,
checks if updates are allowed (is force set, is the udpate
fast-forward?), sends selected ref updates, and then a pack containing
the objects the remote needs.