Add delta encoding
This implements serialization of the Delta::Copy and Delta::Insert
objects. The deltified represenation is described in
https://git-scm.com/docs/pack-format#_deltified_representation.

The variable-length integers in Pack::Numbers:VarIntLE get the number of
bits enocded in the first byte parametrized, as its' 4 when describing
object sizes in a pack, and 7 when encoding lengths of source and target
strings in a delta.