Use delta compression when sending and receiving packs
With paths added to RevList, we can sort objects to be sent in a way
that groups the best candidates for compressibility next to each other.

Each object is compared to a few other objects that are likely to be
similar, and if delta compression reduces the size by at least 50%, it
is applied. The best compression of the current sliding window is
retained. A heuristic is applied to discourage long delta chains.

This new functionality is added to the Writer and Reader classes for
packs.