Add capability to store pack on disk and index it
This adds a new class used when receiving a pack, Pack::Indexer. Instead
of unpacking each object and storing it on disk individually, like
Pack::Unpacker does, Indexer does two passes over the pack:

- In the first pass, we look at each object in the pack, and store ID
  and offset for all non-deltified objects. The offsets of deltified
  objects are retained for the second pass. The pack is written to disk
  as part of the first pass.

- In the second pass, the deltified objects are reconstructed,
  recursively.

The result of these two passes is an in-memory index, mapping each
object ID to its pack offset and the CRC32 of its packed bytes. This
index is written to an index file, starting with a fanout table, then
the sorted list of IDs, the CRC32 of each object, the offsets, and the
checksum of the .pack file.

Offsets larger than 31 bits are handled separately.

Storing a pack on disk is controlled by an "unpackLimit" config value
for fetch and receive: if there are fewer objects than that value, they
are unpacked invidually, and else the pack is stored in disk with an
index. By default, this value is 100, but because we can't read the pack
index yet, we only use the indexer if an unpackLimit is set.