Support ofs-delta records
When reading a pack and encountering a delta-compressed object, we had
to recursively expand the base objects along the entire delta chain, and
base objects were identified by their ID. To find the base, we had to do
a lookup for each object.

The ofs-delta record type identifies its base directly by the offset in
the pack instead of its ID, saving a lookup in the index.

This ability is advertised in receive-pack and fetch.